+++ src/main/java/ssafy/lambda/invitation/controller/InvitationController.java
@@ -11,7 +11,7 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-import ssafy.lambda.global.response.Response;
+import ssafy.lambda.global.response.dto.Response;
 import ssafy.lambda.invitation.dto.RequestAcceptInvitationDto;
 import ssafy.lambda.invitation.dto.RequestCreateInvitationDto;
 import ssafy.lambda.invitation.dto.ResponseInvitationDto;
