+++ src/main/java/ssafy/lambda/team/controller/TeamController.java
@@ -14,7 +14,7 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
-import ssafy.lambda.global.response.Response;
+import ssafy.lambda.global.response.dto.Response;
 import ssafy.lambda.member.entity.Member;
 import ssafy.lambda.member.service.MemberService;
 import ssafy.lambda.team.dto.RequestTeamCreateDto;
